refactor: move everything into attribute macro
