+++ crates/bitwarden-error-macro/src/basic/mod.rs
@@ -1,2 +1,1 @@
 pub mod attribute;
-pub mod derive;
